requirements.txt
Django==5.2.4
razorpay==1.4.2
django-import-export==4.0.0
pycryptodome==3.12.0